Fix two filesystem compiler warnings
md5.hpp:199 [-Wnull-pointer-subtraction]
  Pointer alignment check 'data - (md5_byte_t const*)0' is UB in C++.
  Replace with reinterpret_cast<uintptr_t>(data) which achieves the
  same alignment test without involving a null pointer.

files_decompress.cpp:1021 [-Wsign-compare]
  frz.Read() returns FileReader::Size (long), mSize is size_t.
  Same pattern fixed earlier in g_game.cpp — cast Read() result to
  size_t at the comparison site.

+++ src/common/filesystem/source/files_decompress.cpp
@@ -1018,7 +1018,7 @@ bool FCompressedBuffer::Decompress(char* destbuffer)
 		FileReader frz;
 		if (OpenDecompressor(frz, mr, mSize, mMethod))
 		{
-			return frz.Read(destbuffer, mSize) != mSize;
+			return (size_t)frz.Read(destbuffer, mSize) != mSize;
 		}
 	}
 	return false;
